Added KanjiCategories + Refactored enum types
