+++ lib/core/database/database.dart
@@ -29,7 +29,7 @@ class CustomDatabase {
 
     _database = await openDatabase(
       path,
-      version: 4,
+      version: 5,
       singleInstance: true,
       onConfigure: (db) async {
         await db.execute('PRAGMA foreign_keys = ON');
@@ -53,6 +53,7 @@ class CustomDatabase {
             "${KanjiTableFields.dateLastShownReading} INTEGER NOT NULL DEFAULT 0, "
             "${KanjiTableFields.dateLastShownRecognition} INTEGER NOT NULL DEFAULT 0, "
             "${KanjiTableFields.dateLastShownListening} INTEGER NOT NULL DEFAULT 0, "
+            "${KanjiTableFields.categoryField} INTEGER NOT NULL DEFAULT 0, "
             "PRIMARY KEY (${KanjiTableFields.kanjiField}, ${KanjiTableFields.meaningField}, ${KanjiTableFields.pronunciationField}), "
             "FOREIGN KEY (${KanjiTableFields.listNameField}) "
             "REFERENCES ${KanListTableFields.listsTable}(${KanListTableFields.nameField}) "
@@ -82,6 +83,7 @@ class CustomDatabase {
     if (oldVersion <= 1) Migrations.version1to2(db);
     if (oldVersion <= 2) Migrations.version2to3(db);
     if (oldVersion <= 3) Migrations.version3to4(db);
+    if (oldVersion <= 4) Migrations.version4to5(db);
   }
 
   /// Closes up the current database.
