+++ lib/core/database/migrations.dart
@@ -3,6 +3,12 @@ import 'package:kanpractice/core/database/models/kanji.dart';
 import 'package:sqflite/sqflite.dart';
 
 class Migrations {
+  static Future<void> version4to5(Database db) async {
+    await db.rawQuery("ALTER TABLE ${KanjiTableFields.kanjiTable} "
+        "ADD COLUMN ${KanjiTableFields.categoryField} INTEGER NOT NULL "
+        "DEFAULT 0");
+  }
+
   static Future<void> version3to4(Database db) async {
     await db.rawQuery("ALTER TABLE ${KanjiTableFields.kanjiTable} "
         "ADD COLUMN ${KanjiTableFields.dateLastShownWriting} INTEGER NOT NULL "
