+++ lib/core/utils/types/coach_tutorial_parts.dart
@@ -1,0 +1,1 @@
+enum CoachTutorialParts { kanList, details }
