+++ lib/ui/pages/kanji_lists/kanji_lists.dart
@@ -6,9 +6,11 @@ import 'package:kanpractice/core/preferences/store_manager.dart';
 import 'package:kanpractice/core/routing/pages.dart';
 import 'package:kanpractice/core/tutorial/tutorial_manager.dart';
 import 'package:kanpractice/core/utils/GeneralUtils.dart';
+import 'package:kanpractice/core/utils/types/coach_tutorial_parts.dart';
+import 'package:kanpractice/core/utils/types/visualization_mode.dart';
 import 'package:kanpractice/ui/pages/dictionary/arguments.dart';
 import 'package:kanpractice/ui/pages/kanji_lists/bloc/lists_bloc.dart';
-import 'package:kanpractice/ui/pages/kanji_lists/filters.dart';
+import 'package:kanpractice/core/utils/types/filters.dart';
 import 'package:kanpractice/ui/pages/kanji_lists/widgets/TestBottomSheet.dart';
 import 'package:kanpractice/ui/widgets/CreateKanListDialog.dart';
 import 'package:kanpractice/ui/widgets/CustomSearchBar.dart';
